+++ jpa-shop/src/main/java/jpabook/jpashop/domain/OrderItem.java
@@ -1,5 +1,7 @@
 package jpabook.jpashop.domain;
 
+import jpabook.jpashop.domain.item.Item;
+
 import javax.persistence.*;
 
 @Entity
